requirements.txt
streamlit>=1.32.0
cryptography>=42.0.0
bcrypt>=4.1.0
psycopg2-binary>=2.9.0
